Update example to use local image
This adds an example for how one might use a local image to populate the
canvas. In order to support exporting the final canvas, the image must
be initially either from the client, or the image source must be loaded
with an appropriate CORS header or a security exception will be thrown.

References
https://developer.mozilla.org/en-US/docs/Web/HTML/CORS_enabled_image